+++ tidy3d/plugins/invdes/penalty.py
@@ -10,6 +10,7 @@ import pydantic.v1 as pd
 
 from tidy3d.constants import MICROMETER
 from tidy3d.plugins.autograd.invdes import make_erosion_dilation_penalty
+from tidy3d.plugins.autograd.invdes.spacing import GridCoords
 
 from .base import InvdesBaseModel
 
@@ -90,10 +91,21 @@ class ErosionDilationPenalty(AbstractPenalty):
         "using it unless there is a good reason to set it differently.",
     )
 
-    def evaluate(self, x: anp.ndarray, pixel_size: float) -> float:
+    def evaluate(
+        self,
+        x: anp.ndarray,
+        pixel_size: float | tuple[float, ...] | None = None,
+        *,
+        coords: GridCoords | None = None,
+    ) -> float:
         """Evaluate this penalty."""
         penalty_fn = make_erosion_dilation_penalty(
-            self.length_scale, pixel_size, beta=self.beta, eta=self.eta0, delta_eta=self.delta_eta
+            self.length_scale,
+            pixel_size,
+            coords=coords,
+            beta=self.beta,
+            eta=self.eta0,
+            delta_eta=self.delta_eta,
         )
         penalty_unweighted = penalty_fn(x)
         return self.weight * penalty_unweighted