+++ tidy3d/plugins/invdes/validators.py
@@ -33,7 +33,9 @@ def check_pixel_size(sim_field_name: str):
     """make validator to check the pixel size of sim or list of sims in an ``InverseDesign``."""
 
     def check_pixel_size_sim(
-        sim: td.Simulation, pixel_size: float, index: typing.Optional[int] = None
+        sim: td.Simulation,
+        pixel_size: typing.Union[float, tuple[float, float, float]],
+        index: typing.Optional[int] = None,
     ) -> None:
         """Check a pixel size compared to the simulation min wvl in material."""
         if not sim.sources:
@@ -43,7 +45,8 @@ def check_pixel_size(sim_field_name: str):
             )
             return
 
-        if pixel_size > PIXEL_SIZE_WARNING_THRESHOLD * sim.wvl_mat_min:
+        pixel_sizes = (pixel_size,) if isinstance(pixel_size, (int, float)) else tuple(pixel_size)
+        if any(size > PIXEL_SIZE_WARNING_THRESHOLD * sim.wvl_mat_min for size in pixel_sizes):
             sim_string = f"simulations[{index}]" if index else "the simulation"
 
             td.log.warning(